Consolidate API code into the api directory
API types were spread across api.types.ts and the api directory. This consolidates them into a single place
