+++ openfish-webapp/src/api/index.ts
@@ -1,0 +1,6 @@
+export type Result<T> = {
+  results: T[]
+  offset: number
+  limit: number
+  total: number
+}
