+++ openfish-webapp/src/api/species.ts
@@ -1,0 +1,8 @@
+export type Species = {
+  id: number
+  species: string
+  common_name: string
+  images?: Image[]
+}
+
+export type Image = { src: string; attribution: string }
