+++ openfish-webapp/src/api/videostream.ts
@@ -1,0 +1,7 @@
+export type VideoStream = {
+  id: number
+  stream_url: string
+  capturesource: string
+  startTime: string
+  endTime: string
+}
