+++ openfish-webapp/src/webcomponents/capturesource-list.ts
@@ -1,6 +1,7 @@
 import { LitElement, css, html } from 'lit'
 import { customElement, state } from 'lit/decorators.js'
-import type { CaptureSource, Result } from '../utils/api.types.ts'
+import type { Result } from '../api'
+import type { CaptureSource } from '../api/capturesource.ts'
 import { repeat } from 'lit/directives/repeat.js'
 import resetcss from '../styles/reset.css?lit'
 import btncss from '../styles/buttons.css?lit'
