+++ openfish-webapp/src/webcomponents/stream-list.ts
@@ -1,6 +1,8 @@
 import { LitElement, css, html } from 'lit'
 import { customElement, property, state } from 'lit/decorators.js'
-import type { CaptureSource, Result, VideoStream } from '../utils/api.types.ts'
+import type { VideoStream } from '../api/videostream.ts'
+import type { Result } from '../api'
+import type { CaptureSource } from '../api/capturesource.ts'
 import { repeat } from 'lit/directives/repeat.js'
 import resetcss from '../styles/reset.css?lit'
 import { datetimeDifference, formatAsDate, formatDuration } from '../utils/datetime.ts'
